#35 - refactor: SignUpFormData 타입 types 폴더에 따로 관리
타입을 사용하는 파일들 모두 경로 수정
